// src/V2/types.ts
export type IHyperion = V2Type;
// ...
export interface ICommandContext<T = Module>{
    msg: Message;
    channel: TextChannel;
    guild: Guild;
    guildConfig: GuildType;
    member: Member;
    user: User;
    command: Command;
    module: T;
    content: string;
    args: Array<string>;
    dev: boolean;
    admin: boolean;
    override?: boolean;
    permLevel?: number;
    respond?: string;
    silent?: boolean;
    delete?: boolean;
    mod: boolean;
    createMessage(input: MessageContent): void;
}
// ...
export interface IModerationContext{
    user: string;
    member?: Member;
    moderator: string;
    moderationType: string;
    reason?: string;
    length?: number;
    timeGiven: number;
    caseNumber?: number;
    auto: boolean;
    role?: string;
    removedRoles?: Array<string>;
    guild: Guild;
    mid?: string;
    stringLength?: string;
    moderationEnd?: boolean;
    autoEnd: boolean;
    logChannel?: string;
}

export type EmbedResponse = {embed: Partial<Embed>};
export type MixedResponse = string | EmbedResponse | emoteResponse;
export type CommandResponse = Promise<MixedResponse>;
export type Field = {name: string; value: string; inline: boolean};
export type FieldArray = Array<Field>;